@@ -1,6 +1,5 @@
 package com.example.headstart;
 
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
@@ -50,17 +49,9 @@ public class MainEntryActivity extends AppCompatActivity implements View.OnClick
                 .setTitle("Quit Application")
                 .setCancelable(false)
                 .setMessage("Are your sure to exit?")
-                .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        ActivityCompat.finishAffinity(MainEntryActivity.this);
-                    }
-                })
-                .setNegativeButton("No", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialogInterface, int i) {
+                .setPositiveButton("Yes", (dialog, which) -> ActivityCompat.finishAffinity(MainEntryActivity.this))
+                .setNegativeButton("No", (dialogInterface, i) -> {
 
-                    }
                 });
 
         builder.show();
