@@ -3,7 +3,8 @@ package com.example.headstart;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentPagerAdapter;
+import androidx.lifecycle.Lifecycle;
+import androidx.viewpager2.adapter.FragmentStateAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,36 +13,63 @@ import java.util.List;
 /**
  * fragments are stored in this PagerAdapter class
  */
-public class PagerAdapter extends FragmentPagerAdapter {
-
-   // private final int mNumOfTabs;
+public class PagerAdapter extends FragmentStateAdapter {
 
     /**
      * store fragments in a List, this holds fragment
      */
-    private final List<Fragment> mFragmentList = new ArrayList<Fragment>();
+    private final List<Fragment> mFragmentList = new ArrayList<>();
 
-    public PagerAdapter(@NonNull FragmentManager fm, int behavior ){
-        super(fm, behavior);
-       // this.mNumOfTabs = NumOfTabs; , int NumOfTabs)
+    public PagerAdapter(@NonNull FragmentManager fragmentManager, @NonNull Lifecycle lifecycle) {
+        super(fragmentManager, lifecycle);
     }
-
 
     @NonNull
     @Override
-    public Fragment getItem(int position) {
+    public Fragment createFragment(int position) {
         return mFragmentList.get(position);
     }
 
     @Override
-    public int getCount() {
+    public int getItemCount() {
         return mFragmentList.size();
     }
 
     /**
      * @param fragment , method to add fragment to List
      */
-    public void addFragment(Fragment fragment){
+    public void addFragment(Fragment fragment) {
         mFragmentList.add(fragment);
     }
+
+    // private final int mNumOfTabs;
+
+//    /**
+//     * store fragments in a List, this holds fragment
+//     */
+//    private final List<Fragment> mFragmentList = new ArrayList<Fragment>();
+//
+//    public PagerAdapter(@NonNull FragmentManager fm, int behavior ){
+//        super(fm, behavior);
+//       // this.mNumOfTabs = NumOfTabs; , int NumOfTabs)
+//    }
+//
+//
+//    @NonNull
+//    @Override
+//    public Fragment getItem(int position) {
+//        return mFragmentList.get(position);
+//    }
+//
+//    @Override
+//    public int getCount() {
+//        return mFragmentList.size();
+//    }
+//
+//    /**
+//     * @param fragment , method to add fragment to List
+//     */
+//    public void addFragment(Fragment fragment){
+//        mFragmentList.add(fragment);
+//    }
 }
